+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Changed
 
 - Removed hyphenated words as a possibility for `hangman`.
+- Added the number of winners to raffle embeds when there is more than one winner.
 
 ## 4.78.1569 Kazuki